requirements.txt
flask
gunicorn
beautifulsoup4
markdownify
playwright
sentence-transformers
faiss-cpu
python-dotenv
